:sparkles: feat: 백준 1850번 풀이

+++ src/chapter08/boj_1850/Main.java
@@ -9,8 +9,8 @@ public class Main {
         BufferedWriter bw = new BufferedWriter(new OutputStreamWriter(System.out));
 
         StringTokenizer st = new StringTokenizer(br.readLine());
-        long a = Integer.parseInt(st.nextToken());
-        long b = Integer.parseInt(st.nextToken());
+        long a = Long.parseLong(st.nextToken());
+        long b = Long.parseLong(st.nextToken());
 
         long result = gcd(a, b);
         while (result > 0) {
